# Improve cruise config xml save performance (#838):