--- a/config/config-api/src/com/thoughtworks/go/config/ConfigCacheStore.java
+++ b/config/config-api/src/com/thoughtworks/go/config/ConfigCacheStore.java
@@ -1,27 +1,22 @@
-/*************************GO-LICENSE-START*********************************
- * Copyright 2014 ThoughtWorks, Inc.
+/*
+ * Copyright 2016 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *************************GO-LICENSE-END***********************************/
+ */
 
 package com.thoughtworks.go.config;
 
-import com.thoughtworks.go.domain.ConfigErrors;
-
-import java.io.Serializable;
-
-public interface Validatable extends Serializable {
-    void validate(ValidationContext validationContext);
-    ConfigErrors errors();
-    void addError(String fieldName, String message);
+public interface ConfigCacheStore {
+    Object get(String key);
+    void put(String key, Object obj);
 }