--- a/config/config-api/src/com/thoughtworks/go/config/GoConfigHolder.java
+++ b/config/config-api/src/com/thoughtworks/go/config/GoConfigHolder.java
@@ -16,7 +16,9 @@
 
 package com.thoughtworks.go.config;
 
-public class GoConfigHolder {
+import java.io.Serializable;
+
+public class GoConfigHolder implements Serializable {
     public final CruiseConfig config;
     public final CruiseConfig configForEdit;