--- a/config/config-server/src/com/thoughtworks/go/config/GoConfigMigration.java
+++ b/config/config-server/src/com/thoughtworks/go/config/GoConfigMigration.java
@@ -145,7 +145,7 @@ public class GoConfigMigration {
 
     private GoConfigHolder reloadedConfig(ByteArrayOutputStream stream, String upgradedXmlString) throws Exception {
         GoConfigHolder configHolder = validateAfterMigrationFinished(upgradedXmlString);
-        new MagicalGoConfigXmlWriter(configCache, registry, metricsProbeService).write(configHolder.configForEdit, stream, false);
+        new MagicalGoConfigXmlWriter(configCache, registry, metricsProbeService).write(configHolder.configForEdit, stream);
         return configHolder;
     }