1. Avoid unnecessary validate and deep clone CruiseConfig object
2. Avoid parsing same config string multiple times when saving cruise config xml config
3. Class#getSimpleName is slow when calling it too many times.

Note: only improves save all cruise config xml content (request: POST /go/admin/config_xml), won't improve partial update pipeline config.

-----------------

#### Benchmark

Environment Setup: see 760c0b0212d2490a0bc45ba2c22ca87f93fbab25 and 091a91326e3f5fd5039f0c587c1c039f8d89ba0a for setup environment, these 2 changes make development server more like production environment, and also disabled most of background job

Test config: I copied from our Mingle team's go configuration, it's a configuration has 15.8k lines, lots of pipelines.

Go server need a current config md5 for save config, so, I manually copied all POST data from browser, and I only can change the copied data once to get test result.

Warm up: about 100 requests to update same config.

Then I manually change configuration a little bit to fire the test request.

Ab command example: ab -v 4 -A admin:badger -T 'application/x-www-form-urlencoded' -p post_data -c 1 -n 1 http://localhost:8153/go/admin/config_xml

Result: about 770ms => 700ms when config changed, 1100ms => 400ms when save same config