@xli 2016-02-04
I spent most of time figure out how to get a performance benchmark for showing improvement. 
I'm not happy with current benchmark strategy, too many manual steps.
But it's the best I can get right now.
Anyone has better idea?