@xli 2016-02-04
missed a commented code. need more time to test and try out different solutions.